# newdocs/05-api-swagger.yaml
openapi: 3.0.3
info:
  title: AquaGuard API
  version: "1.0.0"
  description: |
    # AquaGuard REST API

    API của nền tảng ứng phó và điều phối cứu hộ lũ lụt AquaGuard.
    Tài liệu này mô tả **toàn bộ** endpoint HTTP hiện có của backend.

    ## Quy ước chung

    - Mọi phản hồi đều là JSON và luôn có trường `success` (boolean).
      - Thành công: `{ "success": true, "data": ... }` (một số endpoint trả thêm
        `message`, `group`, `unreadCount`, `sentCount`…).
      - Thất bại: `{ "success": false, "message": "..." }`.
    - Ngoại lệ duy nhất: nhóm `/export/*` trả về **file CSV** (`text/csv`).
    - Múi giờ: mọi mốc thời gian là ISO-8601 có offset (`TIMESTAMPTZ`).
    - Số điện thoại luôn ở dạng chuẩn hoá `+84XXXXXXXXX` (9–10 chữ số sau `+84`).

    ## Xác thực

    Hầu hết endpoint yêu cầu **JWT** ở header:

    ```
    Authorization: Bearer <accessToken>
    ```

    Token nhận được từ `POST /auth/register` hoặc `POST /auth/login`, có hạn
    **7 ngày**, và chỉ chứa `id`, `phone_number`, `role`.

    ## Phân quyền

    Ba vai trò: `citizen`, `rescuer`, `admin`. Mỗi endpoint ghi rõ vai trò được
    phép trong phần mô tả. Sai vai trò → **403**. Thiếu/sai token → **401**.

    Một số thao tác còn ràng buộc thêm ở mức dữ liệu (ví dụ chỉ đội đang phụ
    trách mới hoàn tất được nhiệm vụ, chỉ nhóm trưởng mới đổi được vai trò thành
    viên) — các ràng buộc này nằm ngay trong câu lệnh SQL nên trả **403/404**
    kể cả khi vai trò tổng quát đã đúng.

...

...

...
    | --- | --- |
...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /sensors:
    get:
      tags: [Water Sensors]
      summary: Danh sách cảm biến của tôi
      description: |
        Mỗi thiết bị kèm ảnh chụp số đo mới nhất (`percent`, `level`,
        `lastSeenAt`) nên giao diện chỉ cần một lần gọi. `online` là **tính
        toán tại thời điểm đọc**: đúng khi thiết bị có tin trong 45 giây gần
        nhất. Không bao giờ trả về device key.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/WaterSensor" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/ServerError" }

    post:
      tags: [Water Sensors]
      summary: Ghép một board cảm biến mới — chỉ cứu hộ & quản trị
      description: |
        **Người dân không tạo được cảm biến** (403). Cảm biến là thiết bị do đội
        triển khai: người tạo sở hữu nó, chịu trách nhiệm hiệu chuẩn và là người
        nhận cảnh báo khi nó vượt ngưỡng.

        Trả về `deviceKey` **đúng một lần** — máy chủ chỉ giữ SHA-256 của nó.
        Mất key thì dùng `POST /sensors/{id}/rotate-key` để cấp key khác.

        Mỗi tài khoản ghép tối đa **50** cảm biến.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80, example: "Cảm biến sân sau" }
                latitude: { type: number, format: double }
                longitude: { type: number, format: double }
                address: { type: string }
                calibration:
                  description: Bảng hiệu chuẩn `[[phần_trăm, raw], ...]` (từ calib.json trên board).
                  type: array
                  items:
                    type: array
                    items: { type: integer }
                  example: [[0, 0], [50, 18400], [100, 39000]]
      responses:
        "201":
          description: Đã tạo
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    allOf:
                      - $ref: "#/components/schemas/WaterSensor"
                      - type: object
                        properties:
                          deviceKey: { type: string, example: "aqg_XcJ2..." }
        "400":
          description: Vượt quá số cảm biến cho phép
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Người dân không được tạo cảm biến
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
  /sensors/monitor:
    get:
      tags: [Water Sensors]
      summary: Toàn bộ cảm biến mực nước trong hệ thống — chỉ cứu hộ & quản trị
      description: |
        Nguồn dữ liệu cho mục **Mực nước** ở Trung tâm Giám sát. Trả về mọi
        thiết bị của mọi người dùng, kèm chủ sở hữu và chuỗi số đo 6 giờ gần
        nhất (tối đa 24 điểm mỗi thiết bị) để vẽ đường sparkline — một query
        duy nhất, không gọi thêm lần nào cho từng thiết bị.

        Thứ tự trả về đã xếp theo mức độ khẩn: đang vượt ngưỡng → đang ngập từ
        mức 5 → bình thường → mất kết nối. Người trực mở trang là thấy ngay chỗ
        cần lo trước.

        Khác `GET /sensors` (của chính mình) ở hai điểm: có `owner`, và **không**
        có `keyPrefix` — đó là một mẩu khoá bí mật của người khác.

        `owner.phoneNumber` chỉ xuất hiện với **admin**; cứu hộ chỉ thấy tên.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/WaterSensor"
                        - type: object
                          properties:
                            owner:
                              type: object
                              properties:
                                id: { type: integer, example: 7 }
                                displayName: { type: string, example: "Nguyễn Văn A" }
                                phoneNumber:
                                  type: string
                                  description: Chỉ trả về cho admin.
                                  example: "+84901234567"
                            canManage:
                              type: boolean
                              description: >
                                Người gọi có được sửa/xoá thiết bị này không —
                                server tự chấm (admin: mọi thiết bị; cứu hộ:
                                thiết bị mình tạo). Giao diện chỉ dùng để ẩn/hiện nút.
                            history:
                              description: Số đo 6 giờ gần nhất, cũ trước mới sau.
                              type: array
                              items:
                                type: object
                                properties:
                                  percent: { type: integer, example: 63 }
                                  at: { type: string, format: date-time }
                  summary:
                    type: object
                    properties:
                      total: { type: integer, example: 12 }
                      online: { type: integer, example: 9 }
                      offline: { type: integer, example: 3 }
                      warning:
                        type: integer
                        description: Số thiết bị đang online và vượt ngưỡng của chủ thiết bị.
                        example: 2
                      critical:
                        type: integer
                        description: Số thiết bị đang online và ở mức 9 (đầy).
                        example: 1
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Người dân không được xem thiết bị của người khác
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
  /sensors/{id}:
    put:
      tags: [Water Sensors]
      summary: Cập nhật thiết bị (tên, ngưỡng cảnh báo, vị trí, hiệu chuẩn)
      description: |
        Quản trị sửa được **mọi** thiết bị; cứu hộ chỉ sửa thiết bị mình tạo
        (thiết bị của người khác trả về 404). Người dân bị chặn (403).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                alertThreshold: { type: integer, minimum: 1, maximum: 100 }
                alertEnabled: { type: boolean }
                address: { type: string }
                latitude: { type: number, format: double }
                longitude: { type: number, format: double }
                calibration:
                  type: array
                  items:
                    type: array
                    items: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: "#/components/schemas/WaterSensor" }
        "400":
          description: Dữ liệu không hợp lệ
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Không tìm thấy cảm biến của bạn
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }

    delete:
      tags: [Water Sensors]
      summary: Gỡ thiết bị
      description: |
        Xoá luôn toàn bộ số đo đã lưu của thiết bị (ON DELETE CASCADE).

        **Quản trị gỡ được mọi thiết bị** — kể cả của người khác, để dọn thiết
        bị hỏng hoặc của người đã rời đội. Cứu hộ chỉ gỡ được thiết bị mình tạo
        (của người khác trả về 404). Người dân bị chặn (403).
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Đã gỡ
          content: { application/json: { schema: { $ref: "#/components/schemas/SimpleSuccess" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Không tìm thấy cảm biến của bạn
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "500": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT lấy từ `POST /auth/register` hoặc `POST /auth/login`, hạn 7 ngày.
        Với WebSocket, token được truyền qua query string (`?token=...`) thay vì header.

    deviceKeyAuth:
      type: apiKey
      in: header
      name: X-Device-Key
      description: |
        Key dài hạn của một board cảm biến, cấp khi ghép thiết bị
        (`POST /sensors`). Chỉ dùng cho `POST /sensors/ingest` — board không
        đăng nhập được nên không dùng JWT. Máy chủ chỉ lưu SHA-256 của key.
...
  responses:
    Unauthorized:
      description: Thiếu token, token sai hoặc đã hết hạn
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
          example: { success: false, message: "Invalid token." }
...
    ServerError:
      description: Lỗi máy chủ
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Lỗi server" }
...